# docs/openapi.3.0.yaml
openapi: 3.0.0
info:
  title: FeexSystems Living Intelligence World
  description: |
    Evidence-backed engineering intelligence platform that transforms the FeexSystems GitHub ecosystem into an explorable World Model.
    
    **Core Principles:**
    - Canonical Data + Model Reasoning: Database-backed World Model is authoritative; models interpret and summarize.
    - Graph, Not List: Relationships are first-class entities with evidence trails.
    - Evidence, Not Claims: Facts must have traceable GitHub implementation evidence.
    - Provider-Neutral Intelligence: Model reasoning adapters interchange without breaking contracts.
    - Spatial Meaning: 3D interface communicates engineering topology.
  version: 2.0.0
  contact:
    name: FeexSystems API Support
    url: https://feexsystems.codes
    email: support@feexsystems.codes
  license:
    name: Proprietary
    url: https://feexsystems.codes/license
  x-logo:
    url: https://feexsystems.codes/logo.svg
    altText: FeexSystems Logo
servers:
  - url: https://api.feexsystems.codes
    description: Production Server
  - url: http://localhost:3001
    description: Development Server
  - url: http://localhost:8080
    description: Dev Server with Vite Integration
...
components:
  securitySchemes:
    firebaseJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase JWT token in Authorization header
    sessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Optional session token for request tracking
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for server-to-server communication (future)
    mockAuth:
      type: apiKey
      in: header
      name: X-Mock-User
      description: Mock user ID for development (when USE_MOCK_AUTH=true)
  parameters:
    pageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data (structure varies by endpoint)
        metadata:
          type: object
          description: Pagination, timing, or other metadata
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - type
        - message
        - code
        - timestamp
      properties:
        type:
          type: string
          enum:
            - MISSING_TOKEN
            - INVALID_TOKEN
            - TOKEN_EXPIRED
            - AUTHENTICATION_ERROR
            - AUTHENTICATION_REQUIRED
            - INSUFFICIENT_PERMISSIONS
            - VALIDATION_ERROR
            - VALIDATION_FAILED
            - RATE_LIMIT_EXCEEDED
            - NOT_FOUND_ERROR
            - CONFLICT_ERROR
            - INTERNAL_SERVER_ERROR
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          description: Additional error context
        requestId:
          type: string
          format: uuid

    # User & Auth Schemas
    User:
      type: object
      required:
        - id
        - email
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [USER, ADMIN, SUPER_ADMIN, ANALYST]
        profileImageUrl:
          type: string
          format: uri
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    Session:
      type: object
      required:
        - id
        - userId
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: AUTHENTICATION_REQUIRED
              message: Authentication required
              code: AUTH_REQUIRED
              timestamp: "2026-09-14T12:00:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Returns authenticated user profile from Firebase/Prisma
      operationId: getCurrentUser
      security:
        - firebaseJwt: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/sync-user:
    post:
      tags:
        - Authentication
      summary: Sync Firebase user to database
      description: Synchronize Firebase user profile to Prisma on first login
      operationId: syncUser
      security:
        - firebaseJwt: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                profileImageUrl:
                  type: string
                  format: uri
      responses:
        '200':
          description: User synced successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/sessions:
    get:
      tags:
        - Authentication
      summary: Get user sessions
      description: List active sessions for current user
      operationId: getUserSessions
      security:
        - firebaseJwt: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Sessions list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          sessions:
                            type: array
                            items:
                              $ref: '#/components/schemas/Session'
                          pagination:
                            type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/stats:
    get:
      tags:
        - Authentication
      summary: Get authentication statistics
      description: Returns session count and last login timestamp
      operationId: getAuthStats
      security:
        - firebaseJwt: []
      responses:
        '200':
          description: Auth statistics
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          stats:
                            type: object
                            properties:
                              sessions:
                                type: integer
                              lastLogin:
                                type: string
                                format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...